requirements.txt
pip==22.3
asgiref==3.5.2
Django==4.1.2
sqlparse==0.4.3
python-decouple==3.6
pillow==9.2.0
colorthief==0.2.1
djangorestframework==3.14.0
requests==2.28.1